intel: ssp: Revise receive FIFO draining
The receive FIFO needs to be drained in a different way depending when it
is done.
Draining before start we should be taking the overflow condition into
account and we also know that the DMA is not yet started.

At stop time we know that the DMA is enabled, we need to make sure that
we are not racing with it on the draining.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>